moved stuff around
Ended up making a src and test directory to hold my resources ... build was getting to be a problem and this was my only solution that I could find.
