+++ test/UserModel.test.js
@@ -1,5 +1,5 @@
-import logger from './utils/Logger'
-import UserModel from './UserModel'
+import logger from '../src/utils/Logger'
+import UserModel from '../src/UserModel'
 
 const MongodbMemoryServer = require('mongodb-memory-server')
 const mongoServer = new MongodbMemoryServer.MongoMemoryServer()
